Change the way properties on document.body are set while loading canvas.
In the default setup, when no width and height options are passed and
`createGL` is called on page load, it throws an error if the script is
included before the body element, since `document.body.style` is null.
The function already accounts for this fact while appending the canvas
to the page, but not while setting `style.margin = '0px'`. This PR
changes that.